requirements.txt
cfscrape<=2.1.1
certifi<=2021.10.8
dnspython>=2.2.0
requests==2.27.1
impacket==0.9.23
psutil>=5.9.0
icmplib>=2.1.1
git+https://github.com/MHProDev/PyRoxy.git
yarl~=1.7.2
google-cloud-logging==3.0.0
